requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]==20.7

# T-Bank (Tinkoff) Invest API
tinkoff-investments==0.2.0b51

# База данных (async)
sqlalchemy==2.0.23
aiosqlite==0.19.0         # SQLite для разработки/Railway
asyncpg==0.29.0           # PostgreSQL для Railway (опционально)

# Конфигурация
python-dotenv==1.0.0

# Утилиты
httpx==0.25.2             # HTTP клиент (зависимость telegram)
